requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
peewee==3.14.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.6.0
PyMySQL==1.0.2
pyTelegramBotAPI==3.7.6
quickchart.io==0.1.5
requests==2.25.1
rsa==4.7
telegraph==1.4.1
Telethon==1.19.0
toml==0.10.2
urllib3==1.26.2
